refactor(phase-3): move 6 component modules to @agents/core/component
Move types, manager, pagination, index, clients, client-config from
packages/cli/src/lib/component/ to packages/core/src/component/ and
update all imports across providers, commands, and test files.

Create CLI component/index.ts barrel re-exporting core modules.

Tests: 1882 pass / 50 fail / 1 error (no import regressions)
