+++ packages/cli/src/commands/add.ts
@@ -5,11 +5,11 @@
  * Routes through ComponentManager.add() for all supported types.
  */
 
+import { COMPONENT_TYPES, parseComponentType } from '@agents/core/component/types'
 import { createOutput } from '@agents/core/output'
 import { EXIT } from '@agents/core/types'
 import { defineCommand } from 'citty'
 import { createComponentManager } from '../lib/component/factory'
-import { COMPONENT_TYPES, parseComponentType } from '../lib/component/types'
 import { globalArgs } from './shared-args'
 
 export default defineCommand({
