+++ packages/cli/src/commands/init.ts
@@ -5,16 +5,16 @@
  * Supported: skill (dedicated), persona/hook (template-based). Others: placeholder.
  */
 
-import type { OutputFormatter } from '@agents/core/output'
-import { createOutput } from '@agents/core/output'
-import { EXIT } from '@agents/core/types'
-import { defineCommand } from 'citty'
 import {
   COMPONENT_TYPES,
   type ComponentType,
   getComponentMeta,
   parseComponentType,
-} from '../lib/component/types'
+} from '@agents/core/component/types'
+import type { OutputFormatter } from '@agents/core/output'
+import { createOutput } from '@agents/core/output'
+import { EXIT } from '@agents/core/types'
+import { defineCommand } from 'citty'
 import { globalArgs } from './shared-args'
 
 // ---------------------------------------------------------------------------
