+++ packages/cli/src/lib/component/factory.ts
@@ -1,4 +1,4 @@
-import { ComponentManager } from './manager'
+import { ComponentManager } from '@agents/core/component/manager'
 import { LocalAgentProvider } from './provider-agent'
 import { LocalCommandProvider } from './provider-command'
 import { LocalProvider } from './provider-local'
