+++ packages/cli/src/lib/component/provider-agent.ts
@@ -10,9 +10,7 @@
 import { existsSync } from 'node:fs'
 import { readdir, readFile } from 'node:fs/promises'
 import { basename, join, relative } from 'node:path'
-import { CliError, err, ok, type Result } from '@agents/core/types'
-import { parseFrontmatter } from '../chunker'
-import { clampLimit, clampPage, emptyPage, paginateArray } from './pagination'
+import { clampLimit, clampPage, emptyPage, paginateArray } from '@agents/core/component/pagination'
 import type {
   Component,
   ComponentAddOptions,
@@ -24,7 +22,9 @@ import type {
   PublishResult,
   RemoveResult,
   SearchParams,
-} from './types'
+} from '@agents/core/component/types'
+import { CliError, err, ok, type Result } from '@agents/core/types'
+import { parseFrontmatter } from '../chunker'
 
 const SUPPORTED_TYPES: readonly ComponentType[] = ['agent'] as const
