+++ packages/cli/src/lib/component/provider-command.ts
@@ -16,9 +16,7 @@
 import { existsSync } from 'node:fs'
 import { readdir, readFile } from 'node:fs/promises'
 import { join, relative } from 'node:path'
-import { CliError, err, ok, type Result } from '@agents/core/types'
-import { parseFrontmatter } from '../chunker'
-import { clampLimit, clampPage, emptyPage, paginateArray } from './pagination'
+import { clampLimit, clampPage, emptyPage, paginateArray } from '@agents/core/component/pagination'
 import type {
   Component,
   ComponentAddOptions,
@@ -30,7 +28,9 @@ import type {
   PublishResult,
   RemoveResult,
   SearchParams,
-} from './types'
+} from '@agents/core/component/types'
+import { CliError, err, ok, type Result } from '@agents/core/types'
+import { parseFrontmatter } from '../chunker'
 
 const SUPPORTED_TYPES: readonly ComponentType[] = ['command']
