+++ packages/cli/src/lib/component/provider-plugin.ts
@@ -12,9 +12,7 @@
 import { existsSync } from 'node:fs'
 import { readdir } from 'node:fs/promises'
 import { join, relative } from 'node:path'
-import { CliError, err, ok, type Result } from '@agents/core/types'
-import { readPluginManifest } from '../manifest'
-import { clampLimit, clampPage, emptyPage, paginateArray } from './pagination'
+import { clampLimit, clampPage, emptyPage, paginateArray } from '@agents/core/component/pagination'
 import type {
   Component,
   ComponentAddOptions,
@@ -26,7 +24,9 @@ import type {
   PublishResult,
   RemoveResult,
   SearchParams,
-} from './types'
+} from '@agents/core/component/types'
+import { CliError, err, ok, type Result } from '@agents/core/types'
+import { readPluginManifest } from '../manifest'
 
 const SUPPORTED_TYPES: readonly ComponentType[] = ['plugin'] as const
