+++ packages/cli/src/lib/component/provider-smithery.ts
@@ -7,8 +7,7 @@
  * the user has discovered a server via search.
  */
 
-import { CliError, err, ok, type Result } from '@agents/core/types'
-import { clampLimit, clampPage, emptyPage } from './pagination'
+import { clampLimit, clampPage, emptyPage } from '@agents/core/component/pagination'
 import type {
   Component,
   ComponentAddOptions,
@@ -21,7 +20,8 @@ import type {
   PublishResult,
   RemoveResult,
   SearchParams,
-} from './types'
+} from '@agents/core/component/types'
+import { CliError, err, ok, type Result } from '@agents/core/types'
 
 const SMITHERY_API_BASE = 'https://registry.smithery.ai/api/v1'
 const TIMEOUT_MS = 3_000
