+++ packages/cli/src/lib/component/smithery-publish.ts
@@ -10,8 +10,8 @@
  * the configured timeout elapses.
  */
 
+import type { PublishResult } from '@agents/core/component/types'
 import { CliError, err, ok, type Result } from '@agents/core/types'
-import type { PublishResult } from './types'
 
 const SMITHERY_API_BASE = 'https://registry.smithery.ai/api/v1'
 const POLL_INTERVAL_MS = 2_000
