+++ packages/cli/src/lib/init-component.ts
@@ -7,6 +7,7 @@
  */
 
 import { join, resolve } from 'node:path'
+import { type ComponentType, getComponentMeta } from '@agents/core/component/types'
 import {
   ensureDir,
   listDirectory,
@@ -16,7 +17,6 @@ import {
 } from '@agents/core/file-io'
 import { currentDir } from '@agents/core/runtime'
 import { CliError, err, ok, type Result } from '@agents/core/types'
-import { type ComponentType, getComponentMeta } from './component/types'
 
 const PROJECT_ROOT = resolve(currentDir(import.meta), '../../../..')
