+++ packages/core/src/component/client-config.ts
