+++ packages/core/src/component/manager.ts
