+++ packages/core/src/component/pagination.ts
