+++ packages/core/src/component/types.ts
